requirements.txt
# Core Physics & Geometry
genesis-world>=0.3.0       # Universal Physics Engine
build123d>=0.5.0           # Programmatic CAD Kernel
numpy>=1.24.0
scipy>=1.10.0

# AI & Orchestration
langgraph                  # Agent State Machine
langchain                  # Framework
langchain-google-genai     # Gemini 3 Adapter
pydantic>=2.0

# Visualization & UI
streamlit>=1.32.0
stpyvista                  # 3D Mesh Rendering for Streamlit
vtk                        # Visualization Toolkit
panel                      # Interactive Dashboards
watchdog                   # File monitoring
